backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
yfinance==0.2.32
pandas==2.1.3
numpy==1.26.2
groq==0.12.0
python-dotenv==1.0.0
